deploy/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
jinja2
python-multipart
